S1.6.3 Check/update schema - Added employee status (populated with values) S1.6.7 Use DTOs - register user not working correctly - User registration works as expected now

+++ BackEnd/src/main/java/com/unosquare/admin_core/back_end/controller/ClientController.java
@@ -79,7 +79,7 @@ public class ClientController {
 
     @GetMapping(value = "/findByClientStatus/{clientStatusId}", produces = MediaType.APPLICATION_JSON_VALUE)
     @ResponseStatus(HttpStatus.OK)
-    public List<ClientDto> findByClientStatus(@PathVariable("clientStatusId") short clientStatusId) {
+    public List<ClientDto> findByClientStatus(@PathVariable("clientStatusId") int clientStatusId) {
         return mapClientsToDtos(clientService.findByClientStatus(ClientStatus.fromId(clientStatusId)));
     }
