+++ BackEnd/src/main/java/com/unosquare/admin_core/back_end/dto/HolidayDto.java
@@ -15,7 +15,7 @@ public class HolidayDto {
 
     private EmployeeDto employee;
 
-    private short holidayStatusId;
+    private int holidayStatusId;
     private String holidayStatusDescription;
 
     private boolean isHalfDay;
