+++ BackEnd/src/main/java/com/unosquare/admin_core/back_end/enums/Country.java
@@ -5,18 +5,18 @@ import com.fasterxml.jackson.annotation.JsonFormat;
 @JsonFormat(shape = JsonFormat.Shape.OBJECT)
 public enum Country {
 
-    NORTHERN_IRELAND((short) 1, "Northern Ireland"),
-    MEXICO((short) 2, "Mexico");
+    NORTHERN_IRELAND(1, "Northern Ireland"),
+    MEXICO(2, "Mexico");
 
-    private final short countryId;
+    private final int countryId;
     private final String description;
 
-    Country(short countryId, String description) {
+    Country(int countryId, String description) {
         this.countryId = countryId;
         this.description = description;
     }
 
-    public short getCountryId() {
+    public int getCountryId() {
         return countryId;
     }
 
@@ -31,7 +31,7 @@ public enum Country {
                 countryId, description);
     }
 
-    public static Country fromId(short countryId) {
+    public static Country fromId(int countryId) {
 
         for (Country status : Country.values()) {
             if (status.countryId == countryId) {
